# packages/api/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"GitConfig": {
				"properties": {
					"uri": {
						"type": "string"
					},
					"branch": {
						"type": "string"
					},
					"encodedToken": {
						"type": "string"
					}
				},
				"required": [
					"uri",
					"branch",
					"encodedToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostBody": {
				"properties": {
					"source": {
						"$ref": "#/components/schemas/GitConfig"
					},
					"destination": {
						"$ref": "#/components/schemas/GitConfig"
					},
					"template": {
						"properties": {
							"data": {}
						},
						"required": [
							"data"
						],
						"type": "object"
					}
				},
				"required": [
					"source",
					"destination",
					"template"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HealthInfo": {
				"properties": {
					"name": {
						"type": "string"
					},
					"version": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"Happy",
							"Sad"
						]
					},
					"phoneNumbers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"phoneNumbers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.email-or-name-or-phoneNumbers_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phoneNumbers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"name",
					"phoneNumbers"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserCreationParams": {
				"$ref": "#/components/schemas/Pick_User.email-or-name-or-phoneNumbers_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@code-rookie/api",
		"version": "0.0.1",
		"license": {
			"name": "MIT License"
		},
		"contact": {
			"name": "Alan S. Ferreira"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/git/render": {
			"post": {
				"operationId": "Render",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostBody"
							}
						}
					}
				}
			}
		},
		"/health": {
			"get": {
				"operationId": "GetHealth",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HealthInfo"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}